requirements.txt
openai
redis
pydantic
tenacity
pyyaml
requests
numpy
reportlab
